RavenDB-26838 CDC Sink: split TaskTypeExtensions into its own file and rename to GetErrorsEndpoint/GetTaskCategoryFromEtlType (review)
