+++ test/Tests.Infrastructure/RavenTestBase.Etl.cs
@@ -246,7 +246,7 @@ namespace FastTests
             {
                 var database = await _parent.GetDatabase(databaseName);
 
-                var errors = database.TaskErrorsStorage.ReadItemErrorsOfTask(TaskTypeExtensions.FromEtlType(config.EtlType), $"{config.Name}/{config.Transforms.First().Name}");
+                var errors = database.TaskErrorsStorage.ReadItemErrorsOfTask(TaskTypeExtensions.GetTaskCategoryFromEtlType(config.EtlType), $"{config.Name}/{config.Transforms.First().Name}");
 
                 return errors.Where(error => error.Step == (int)TaskErrorStep.Load);
             }
@@ -255,7 +255,7 @@ namespace FastTests
             {
                 var database = await _parent.GetDatabase(databaseName);
 
-                var errors = database.TaskErrorsStorage.ReadProcessErrorsOfTask(TaskTypeExtensions.FromEtlType(config.EtlType), $"{config.Name}/{config.Transforms.First().Name}");
+                var errors = database.TaskErrorsStorage.ReadProcessErrorsOfTask(TaskTypeExtensions.GetTaskCategoryFromEtlType(config.EtlType), $"{config.Name}/{config.Transforms.First().Name}");
 
                 return errors.Where(error => error.Step == (int)TaskErrorStep.Load);
             }
@@ -264,7 +264,7 @@ namespace FastTests
             {
                 var database = await _parent.GetDatabase(databaseName);
 
-                var errors = database.TaskErrorsStorage.ReadItemErrorsOfTask(TaskTypeExtensions.FromEtlType(config.EtlType), $"{config.Name}/{config.Transforms.First().Name}");
+                var errors = database.TaskErrorsStorage.ReadItemErrorsOfTask(TaskTypeExtensions.GetTaskCategoryFromEtlType(config.EtlType), $"{config.Name}/{config.Transforms.First().Name}");
 
                 return errors.Where(error => error.Step == (int)TaskErrorStep.Transformation);
             }
